Bestandsveränderungen etc.

+++ Wirtschaftsplanspiel/src/Client/Application/CompanyController.java
@@ -238,8 +238,8 @@ public abstract class CompanyController {
 				return 0;
 			if (missingUnits == 0)
 				return 0;
-				int missMach = (int) ((missingUnits / capacity));
-
+				int missMach = (missingUnits / capacity);
+				
 				return missMach;
 		}
